Fix the validation errors for categories
